% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_png.R
\name{get_png}
\alias{get_png}
\title{Import a PNG image for use in ggplot.}
\usage{
get_png(filename)
}
\arguments{
\item{filename}{the location of the image.}
}
\value{
a png object.
}
\description{
Import a PNG image for use in ggplot.
}
\examples{
logo <- get_png(RCurl::getURLContent("https://cdn.freebiesupply.com/logos/large/2x/tottenham#'-hotspur-logo-png-transparent.png"))
}
