% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitch_us.R
\name{pitch_us}
\alias{pitch_us}
\title{Plot a fill-less half pitch for an understat shotmap.}
\usage{
pitch_us(filler = "#2f4f4f")
}
\arguments{
\item{filler}{the colour you want your lines to be.}
}
\value{
a ggplot annotation half pitch.
}
\description{
Plot a fill-less half pitch for an understat shotmap.
}
\examples{
library(understatr)
library(ggplot2)

data <- get_player_shots(629)

ggplot(data = data, aes(x = X, y = Y, size = xG,
                        colour = result,
                        shape = shotType, fill = result),
       alpha = 0.7) +
  pitch_us() +
  geom_point()
}
